refactor: a lot things has been refactored and fixed

+++ notifications.go
@@ -1,6 +1,6 @@
 package main
 
-func putNotification(note string, x uint8, y uint8) {
+func putNotification(note string, x int16, y int16) {
 	for _, i := range note {
 		world[convertToRealPosition(x, y)] = byte(i)
 		x++
